refactor(step2): production 패키지의 profanities 의존성 제거
