+++ src/main/java/kitchenpos/menus/tobe/domain/Menu.java
@@ -2,11 +2,7 @@ package kitchenpos.menus.tobe.domain;
 
 import kitchenpos.menus.tobe.domain.exception.InvalidMenuArgumentNullPointException;
 import kitchenpos.menus.tobe.domain.exception.InvalidMenuPriceException;
-import kitchenpos.menus.tobe.domain.vo.MenuGroupId;
-import kitchenpos.menus.tobe.domain.vo.MenuId;
-import kitchenpos.menus.tobe.domain.vo.MenuName;
-import kitchenpos.menus.tobe.domain.vo.MenuPrice;
-import kitchenpos.products.tobe.domain.vo.Profanities;
+import kitchenpos.menus.tobe.domain.vo.*;
 
 import java.util.List;
 import java.util.Objects;
