+++ src/main/java/kitchenpos/menus/tobe/domain/vo/EmptyProfanities.java
@@ -1,0 +1,9 @@
+package kitchenpos.menus.tobe.domain.vo;
+
+public class EmptyProfanities implements Profanities {
+
+    @Override
+    public boolean contains(final String name) {
+        return false;
+    }
+}
