+++ src/main/java/kitchenpos/menus/tobe/domain/vo/Profanities.java
@@ -1,0 +1,5 @@
+package kitchenpos.menus.tobe.domain.vo;
+
+public interface Profanities {
+    boolean contains(String name);
+}
