+++ src/test/java/kitchenpos/menus/tobe/domain/MenuTest.java
@@ -1,9 +1,8 @@
 package kitchenpos.menus.tobe.domain;
 
 import kitchenpos.menus.tobe.domain.exception.*;
-import kitchenpos.products.tobe.domain.FakeProfanities;
-import kitchenpos.products.tobe.domain.vo.EmptyProfanities;
-import kitchenpos.products.tobe.domain.vo.Profanities;
+import kitchenpos.menus.tobe.domain.vo.EmptyProfanities;
+import kitchenpos.menus.tobe.domain.vo.Profanities;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
